Decided to move each Game to its own file as a data object with all fields filled since it was getting confusing to scroll back and forth and between files to figure out rules. Moved Family sealed classes into Games.
