+++ app/src/androidTest/java/com/heyzeusv/solitaire/ui/tools/StatsMenuTest.kt
@@ -11,8 +11,11 @@ import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.heyzeusv.solitaire.R
+import com.heyzeusv.solitaire.ui.board.games.Alaska
+import com.heyzeusv.solitaire.ui.board.games.AustralianPatience
+import com.heyzeusv.solitaire.ui.board.games.Games
+import com.heyzeusv.solitaire.ui.board.games.KlondikeTurnOne
 import com.heyzeusv.solitaire.ui.toolbar.menu.StatsMenu
-import com.heyzeusv.solitaire.util.Games
 import com.heyzeusv.solitaire.util.getStatsDefaultInstance
 import com.heyzeusv.solitaire.util.onNodeWithTextId
 import com.heyzeusv.solitaire.util.theme.SolitaireTheme
@@ -45,9 +48,9 @@ class StatsMenuTest {
             setContent {
                 SolitaireTheme {
                     StatsMenu(
-                        selectedGame = Games.KLONDIKE_TURN_ONE,
-                        updateSelectedGame = { },
-                        stats = gameOneStats
+                        selectedGame = KlondikeTurnOne,
+                        updateSelectedGameStats = { },
+                        selectedGameStats = gameOneStats
                     ) { }
                 }
             }
@@ -63,16 +66,16 @@ class StatsMenuTest {
             setContent {
                 SolitaireTheme {
                     StatsMenu(
-                        selectedGame = Games.KLONDIKE_TURN_ONE,
-                        updateSelectedGame = { },
-                        stats = gameOneStats
+                        selectedGame = KlondikeTurnOne,
+                        updateSelectedGameStats = { },
+                        selectedGameStats = gameOneStats
                     ) { }
                 }
             }
 
             onNode(hasTestTag("DropDownMenu")).performClick()
-            Games.entries.forEach { game ->
-                onNode(hasTestTag("DropDownMenu Item ${game.name}")).assertIsDisplayed()
+            Games.orderedSubclasses.forEach {
+                onNode(hasTestTag("DropDownMenu Item")).assertIsDisplayed()
             }
         }
     }
@@ -82,11 +85,11 @@ class StatsMenuTest {
         composeRule.apply {
             setContent {
                 SolitaireTheme {
-                    var selectedGame by remember { mutableStateOf(Games.KLONDIKE_TURN_ONE) }
+                    var selectedGame by remember { mutableStateOf<Games>(KlondikeTurnOne) }
                     var stats by remember { mutableStateOf(gameOneStats) }
                     StatsMenu(
                         selectedGame = selectedGame,
-                        updateSelectedGame = {
+                        updateSelectedGameStats = {
                             selectedGame = it
                             stats = if (stats == gameOneStats) {
                                 gameTwoStats
@@ -94,21 +97,21 @@ class StatsMenuTest {
                                 gameOneStats
                             }
                         },
-                        stats = stats
+                        selectedGameStats = stats
                     ) { }
                 }
             }
 
             // switch to Australian Patience and check stats
             onNode(hasTestTag("DropDownMenu")).performClick()
-            onNode(hasTestTag("DropDownMenu Item ${Games.AUSTRALIAN_PATIENCE}")).performClick()
-            onNodeWithTextId(Games.AUSTRALIAN_PATIENCE.nameId).assertIsDisplayed()
+            onNode(hasTestTag("DropDownMenu Item")).performClick()
+            onNodeWithTextId(AustralianPatience.nameId).assertIsDisplayed()
             gameTwoStatsDisplayed()
 
             // switch to Alaska and check stats
             onNode(hasTestTag("DropDownMenu")).performClick()
-            onNode(hasTestTag("DropDownMenu Item ${Games.ALASKA}")).performClick()
-            onNodeWithTextId(Games.ALASKA.nameId).assertIsDisplayed()
+            onNode(hasTestTag("DropDownMenu Item")).performClick()
+            onNodeWithTextId(Alaska.nameId).assertIsDisplayed()
             gameOneStatsDisplayed()
         }
     }
