+++ app/src/main/java/com/heyzeusv/solitaire/ui/board/games/KlondikeTurnThree.kt
@@ -1,0 +1,50 @@
+package com.heyzeusv.solitaire.ui.board.games
+
+import com.heyzeusv.solitaire.Game
+import com.heyzeusv.solitaire.R
+import com.heyzeusv.solitaire.data.Card
+import com.heyzeusv.solitaire.data.pile.Stock
+import com.heyzeusv.solitaire.data.pile.Tableau
+import com.heyzeusv.solitaire.util.DrawAmount
+import com.heyzeusv.solitaire.util.MaxScore
+import com.heyzeusv.solitaire.util.Redeals
+import com.heyzeusv.solitaire.util.ResetFaceUpAmount
+
+data object KlondikeTurnThree : Games.KlondikeFamily() {
+    /**
+     *  [GameInfo]
+     */
+    override val nameId: Int = R.string.games_klondike_turn_three
+    override val familyId: Int = R.string.games_family_klondike
+    override val previewId: Int = R.drawable.preview_klondike_turn_three
+    override val dataStoreEnum: Game = Game.GAME_KLONDIKETURNTHREE
+
+    /**
+     *  [GameRules]
+     */
+    override val baseDeck: List<Card> = List(52) { Card(it % 13, getSuit(it)) }
+    override val resetFaceUpAmount: ResetFaceUpAmount = ResetFaceUpAmount.One
+    override val drawAmount: DrawAmount = DrawAmount.Three
+    override val redeals: Redeals = Redeals.Unlimited
+    override val maxScore: MaxScore = MaxScore.ONE_DECK
+    override val anyCardCanStartPile: Boolean = false
+
+    override fun autocompleteTableauCheck(tableauList: List<Tableau>): Boolean {
+        tableauList.forEach { if (it.faceDownExists()) return false }
+        return true
+    }
+
+    override fun resetTableau(tableauList: List<Tableau>, stock: Stock) {
+        tableauList.forEachIndexed { index, tableau ->
+            val cards = List(index + 1) { stock.remove() }
+            tableau.reset(resetFlipCard(cards, resetFaceUpAmount))
+        }
+    }
+
+    override fun canAddToTableauRule(tableau: Tableau, cardsToAdd: List<Card>): Boolean {
+        val tLast = tableau.truePile.last()
+        val cFirst = cardsToAdd.first()
+
+        return cFirst.suit.color != tLast.suit.color && cFirst.value == tLast.value - 1
+    }
+}
